+++ db/migrate/20230330093255_create_dailies.rb
@@ -2,8 +2,8 @@ class CreateDailies < ActiveRecord::Migration[6.0]
   def change
     create_table :dailies do |t|
       t.date        :input_day, null: false
-      t.integer     :weight,    null: false
-      t.integer     :fat,       null: false
+      t.float       :weight,    null: false
+      t.float       :fat,       null: false
       t.references  :user,      null: false, foreign_key: true
       t.timestamps
     end
